+++ CHANGELOG.md
@@ -31,6 +31,9 @@ The main work (all changes without a GitHub username in brackets in the below li
     - Fix: Ensure the incoming order of attribute changes is preserved when processing them even though no one should rely on any order
     - Fix: Better handle errors when the BLE connection is disconnected during a write action
 
+- @matter/react-native
+    - Breaking: We updated to @react-native-async-storage/async-storage v3. A v2-compatible class is available. See the package readme.
+
 - @project-chip/matter.js
     - Adjustment: The "Waiting for device discovery" node state is now bound to the availability of IP announcements from MDNS